devtools/server/tests/mochitest/test_styles-matched.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript;version=1.8" src="inspector-helpers.js"></script>
  <script type="application/javascript;version=1.8">
const inspector = require("devtools/server/actors/inspector");
const {CssLogic} = require("devtools/shared/styleinspector/css-logic");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
}

var gWalker = null;
var gStyles = null;
var gClient = null;

addTest(function setup() {
  let url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    gInspectee = doc;
    let {InspectorFront} = require("devtools/server/actors/inspector");
    let inspector = InspectorFront(client, tab);
    promiseDone(inspector.getWalker().then(walker => {
      ok(walker, "getWalker() should return an actor.");
      gClient = client;
      gWalker = walker;
      return inspector.getPageStyle();
    }).then(styles => {
      gStyles = styles;
    }).then(runNextTest));
  });
});

addTest(function testMatchedStyles() {
  promiseDone(gWalker.querySelector(gWalker.rootNode, "#matched-test-node").then(node => {
    return gStyles.getMatchedSelectors(node, "font-size", {});
  }).then(matched => {
    is(matched[0].sourceText, "this.style", "First match comes from the element style");
    is(matched[0].selector, "@element.style", "Element style has a special selector");
    is(matched[0].value, "10px", "First match has the expected value");
    is(matched[0].status, CssLogic.STATUS.BEST, "First match is the best match")
    is(matched[0].rule.type, 100, "First match is an element style");
    is(matched[0].rule.href, gInspectee.defaultView.location.href, "Node style comes from this document")

    is(matched[1].sourceText, ".inheritable-rule", "Second match comes from a rule");
    is(matched[1].selector, ".inheritable-rule", "Second style has a selector");
    is(matched[1].value, "15px", "Second match has the expected value");
    is(matched[1].status, CssLogic.STATUS.PARENT_MATCH, "Second match is from the parent")
    is(matched[1].rule.parentStyleSheet.href, null, "Inline stylesheet shouldn't have an href");
    is(matched[1].rule.parentStyleSheet.nodeHref, gInspectee.defaultView.location.href, "Inline stylesheet's nodeHref should match the current document");
    ok(!matched[1].rule.parentStyleSheet.system, "Inline stylesheet shouldn't be a system stylesheet.");
  }).then(runNextTest));
});

addTest(function testSystemStyles() {
  let testNode = null;

  promiseDone(gWalker.querySelector(gWalker.rootNode, "#matched-test-node").then(node => {
    testNode = node;
    return gStyles.getMatchedSelectors(testNode, "display", { filter: "user" });
  }).then(matched => {
    is(matched.length, 0, "No user selectors apply to this rule.");
    return gStyles.getMatchedSelectors(testNode, "display", { filter: "ua" });
  }).then(matched => {
    is(matched[0].selector, "div", "Should match system div selector");
    is(matched[0].value, "block");
  }).then(runNextTest));
});

addTest(function cleanup() {
  delete gStyles;
  delete gWalker;
  delete gClient;
  runNextTest();
});

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-styles-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
